Modified a file

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -11,7 +11,7 @@
 */
 size_t print_list(const list_t *h)
 {
-int n = 0;
+size_t n = 0;
 while (h != NULL)
 {
 if (h->str == NULL)
@@ -27,4 +27,4 @@ h = h->next;
 n++;
 }
 return (n);
-}
+}
